+++ model/legacy-private/src/main/java/org/keycloak/services/scheduled/StartAutoUpdatedScheduledTasks.java
@@ -45,7 +45,7 @@ public class StartAutoUpdatedScheduledTasks implements ScheduledTask {
                 OpenIdFederationIdPExpirationTask task = new OpenIdFederationIdPExpirationTask(identityProvider.getAlias(), realm.getId());
                 long expiration = Long.valueOf(identityProvider.getConfig().get(OIDCConfigAttributes.EXPIRATION_TIME)) * 1000 - Time.currentTimeMillis();
                 ClusterAwareScheduledTaskRunner taskRunner = new ClusterAwareScheduledTaskRunner(session.getKeycloakSessionFactory(), task, expiration);
-                timer.schedule(taskRunner, expiration, "OpenIdFederationIdPExpirationTask_" + identityProvider.getAlias());
+                timer.scheduleOnce(taskRunner, expiration > 60 * 1000 ? expiration : 60 * 1000, "OpenIdFederationIdPExpirationTask_" + identityProvider.getAlias());
             });
             session.clients().getClientsStream(realm).forEach(clientModel -> {
                 if ("saml".equals(clientModel.getProtocol()) && clientModel.getAttributes() != null && Boolean.valueOf(clientModel.getAttributes().get(SAML_AUTO_UPDATED))) {