+++ services/src/main/java/org/keycloak/services/clientregistration/openid_federation/OpenIdFederationClientRegistrationService.java
@@ -1,5 +1,6 @@
 package org.keycloak.services.clientregistration.openid_federation;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.core.Context;
@@ -27,10 +28,13 @@ import org.keycloak.services.Urls;
 import org.keycloak.services.clientregistration.AbstractClientRegistrationProvider;
 import org.keycloak.services.clientregistration.oidc.DescriptionConverter;
 import org.keycloak.urls.UrlType;
+import org.keycloak.util.JsonSerialization;
 import org.keycloak.util.TokenUtil;
 import org.keycloak.utils.OpenIdFederationTrustChainProcessorFactory;
 
+import java.io.IOException;
 import java.net.URI;
+import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -65,29 +69,41 @@ public class OpenIdFederationClientRegistrationService extends AbstractClientReg
             trustChainProcessor.validationRules(statement, true);
 
             logger.info("starting validating trust chains");
-            TrustChainResolution validChain = trustChainProcessor.constructTrustChains(statement, realm.getOpenIdFederations().stream().map(OpenIdFederationConfig::getTrustAnchor).collect(Collectors.toSet()), true);
-            if (validChain == null) {
-                throw new ErrorResponseException(Errors.INVALID_TRUST_ANCHOR, "No trusted trust anchor could be found", Response.Status.NOT_FOUND);
-            }
-            RPMetadata rPMetadata = (RPMetadata) validChain.getMetadataAfterPolicies();
-
-            ClientRepresentation client = createOrUpdateClient(statement, rPMetadata);
-            URI uri = session.getContext().getUri().getAbsolutePathBuilder().path(client.getClientId()).build();
-            RPMetadata rPMetadataResponse = DescriptionConverter.toExternalResponse(session, client, uri, RPMetadata.class, rPMetadata.getScope()!=null && rPMetadata.getScope().contains(OAuth2Constants.SCOPE_OPENID));
-            event.detail(Details.GRANTED_CLIENT, rPMetadataResponse.getScope());
-            rPMetadataResponse.setClientIdIssuedAt(Time.currentTime());
-
-            rPMetadataResponse.setClientRegistrationTypes(Stream.of(ClientRegistrationTypeEnum.EXPLICIT.getValue()).collect(Collectors.toList()));
-            rPMetadataResponse.setCommonMetadata(rPMetadata.getCommonMetadata());
-            EntityStatementExplicitResponse responseStatement = new EntityStatementExplicitResponse(statement, Urls.realmIssuer(session.getContext().getUri(UrlType.FRONTEND).getBaseUri(), session.getContext().getRealm().getName()), rPMetadataResponse, validChain.getTrustAnchorId(), validChain.getLeafId());
-            responseStatement.type(TokenUtil.EXPLICIT_REGISTRATION_RESPONSE_JWT);
-            String token = session.tokens().encodeForOpenIdFederation(responseStatement);
-            return Response.ok(token).header("Content-Type", TokenUtil.APPLICATION_EXPLICIT_REGISTRATION_RESPONSE_JWT).build();
+            return commonExplicitRegistrationProcess(statement, realm);
 
         } else {
-            // TODO Handle Trust Chain
-            throw new ErrorResponseException("not_implemented", "Trust chain handling is not yet implemented", Response.Status.NOT_IMPLEMENTED);
+            EntityStatement statement = null;
+            try {
+                List<String> trustChains = JsonSerialization.readValue(body, new TypeReference<List<String>>() {});
+                statement = trustChainProcessor.parseAndValidateSelfSigned(trustChains.get(0));
+            } catch (Exception e) {
+                throw new ErrorResponseException(Errors.INVALID_REQUEST, "Body must be Trust Chains", Response.Status.BAD_REQUEST);
+            }
+            trustChainProcessor.validationRules(statement, false);
+            return commonExplicitRegistrationProcess(statement, realm);
         }
+    }
+
+    private Response commonExplicitRegistrationProcess(EntityStatement statement, RealmModel realm){
+        logger.info("starting validating trust chains");
+        TrustChainResolution validChain = trustChainProcessor.constructTrustChains(statement, realm.getOpenIdFederations().stream().map(OpenIdFederationConfig::getTrustAnchor).collect(Collectors.toSet()), true);
+        if (validChain == null) {
+            throw new ErrorResponseException(Errors.INVALID_TRUST_ANCHOR, "No trusted trust anchor could be found", Response.Status.NOT_FOUND);
+        }
+        RPMetadata rPMetadata = (RPMetadata) validChain.getMetadataAfterPolicies();
+
+        ClientRepresentation client = createOrUpdateClient(statement, rPMetadata);
+        URI uri = session.getContext().getUri().getAbsolutePathBuilder().path(client.getClientId()).build();
+        RPMetadata rPMetadataResponse = DescriptionConverter.toExternalResponse(session, client, uri, RPMetadata.class, rPMetadata.getScope()!=null && rPMetadata.getScope().contains(OAuth2Constants.SCOPE_OPENID));
+        event.detail(Details.GRANTED_CLIENT, rPMetadataResponse.getScope());
+        rPMetadataResponse.setClientIdIssuedAt(Time.currentTime());
+
+        rPMetadataResponse.setClientRegistrationTypes(Stream.of(ClientRegistrationTypeEnum.EXPLICIT.getValue()).collect(Collectors.toList()));
+        rPMetadataResponse.setCommonMetadata(rPMetadata.getCommonMetadata());
+        EntityStatementExplicitResponse responseStatement = new EntityStatementExplicitResponse(statement, Urls.realmIssuer(session.getContext().getUri(UrlType.FRONTEND).getBaseUri(), session.getContext().getRealm().getName()), rPMetadataResponse, validChain.getTrustAnchorId(), validChain.getLeafId());
+        responseStatement.type(TokenUtil.EXPLICIT_REGISTRATION_RESPONSE_JWT);
+        String token = session.tokens().encodeForOpenIdFederation(responseStatement);
+        return Response.ok(token).header("Content-Type", TokenUtil.APPLICATION_EXPLICIT_REGISTRATION_RESPONSE_JWT).build();
     }
 
     public ClientRepresentation createOrUpdateClient(EntityStatement statement, RPMetadata rPMetadata){